UI: Add conditional enterprise logo (#4432)
Adds additional 'enterprise' text underneath the 'startup' logo if the
ui is built with a CONSUL_BINARY_TYPE environment variable that doesn't
equal `oss`.

+++ build-support/functions/20-build.sh
@@ -63,12 +63,17 @@ function build_ui {
    then
       commit_hash=$(git rev-parse --short HEAD)
    fi
+   local logo_type="${CONSUL_BINARY_TYPE}"
+   if test "$logo_type" != "oss"
+   then
+     logo_type="enterprise"
+   fi
    
    # make sure we run within the ui dir
    pushd ${ui_dir} > /dev/null
    
    status "Creating the UI Build Container with image: ${image_name} and version '${version}'"
-   local container_id=$(docker create -it -e "CONSUL_GIT_SHA=${commit_hash}" -e "CONSUL_VERSION=${version}" ${image_name})
+   local container_id=$(docker create -it -e "CONSUL_GIT_SHA=${commit_hash}" -e "CONSUL_VERSION=${version}" -e "CONSUL_BINARY_TYPE=${CONSUL_BINARY_TYPE}" ${image_name})
    local ret=$?
    if test $ret -eq 0
    then
@@ -82,7 +87,8 @@ function build_ui {
       ret=$?
       docker rm ${container_id} > /dev/null
    fi
-   
+
+   # Check the version is baked in correctly
    if test ${ret} -eq 0
    then
       local ui_vers=$(ui_version "${1}/ui-v2/dist/index.html")
@@ -90,11 +96,26 @@ function build_ui {
       then
          err "ERROR: UI version mismatch. Expecting: '${version}' found '${ui_vers}'"
          ret=1
-      else
-         rm -rf ${1}/pkg/web_ui/v2
-         mkdir -p ${1}/pkg/web_ui
-         cp -r ${1}/ui-v2/dist ${1}/pkg/web_ui/v2 
       fi
+   fi
+
+   # Check the logo is baked in correctly
+   if test ${ret} -eq 0
+   then
+     local ui_logo_type=$(ui_logo_type "${1}/ui-v2/dist/index.html")
+     if test "${logo_type}" != "${ui_logo_type}"
+     then
+       err "ERROR: UI logo type mismatch. Expecting: '${logo_type}' found '${ui_logo_type}'"
+       ret=1
+     fi
+   fi
+
+   # Copy UI over ready to be packaged into the binary
+   if test ${ret} -eq 0
+   then
+      rm -rf ${1}/pkg/web_ui/v2
+      mkdir -p ${1}/pkg/web_ui
+      cp -r ${1}/ui-v2/dist ${1}/pkg/web_ui/v2 
    fi
    
    popd > /dev/null
